Move `api` code away from `lib`
