+++ src/lib/pages/podcast-episodes/index.ts
@@ -1,2 +1,2 @@
 export { default as PodcastEpisodesPage } from './ui/podcast-episodes-page.svelte';
-export { loadPodcastFromURL } from './lib/load-podcast-from-url';
+export { loadPodcastFromURL } from './api/load-podcast-from-url';
